# Optimized reading rows in a sequential mode, if they are already in the buffer