@vonzshik 2021-01-07
> Thanks for investigating and optimizing this @vonzshik... great to see sequential getting some love :) Do you have the updated stats on sequential vs. non-sequential? I'm curious where the gap is (if at all) after this.
> 
> One thought about this... If I'm reading the code right, your optimization only works if the _entire_ message fits in the buffer. If it doesn't, and the user reads additional data into the buffer, and then does Read to go to the next row, we still invoke ConsumeRowSequential although we don't have to. In other words, unless I'm mistaken, we could update _dataMsgEnd as we read row data, so that the optimization would kick in even for rows that are larger than the buffer. I'm not sure exactly how valuable this would be, what do you think?

Yeah, that's exactly how it works. I did take a look, and that's a very tricky to fix. The problem is, `_dataMsgEnd` is essentially an end position of the row in the buffer. And the only time when we can (and should) update it, is when the buffer is _reset_ (that is, some data was requested from the buffer when it's full).