@vonzshik 2021-01-07
> Do you have the updated stats on sequential vs. non-sequential? I'm curious where the gap is (if at all) after this.

Not sure what you've meant here - do you want me to test some other cases?