@roji 2021-01-07
@vonzshik am just curious if after your changes sequential is still slower than non-sequential, and by how much :) But not urgent or blocking for this PR of course.