@roji 2021-01-07
> I did take a look, and that's a very tricky to fix. The problem is, _dataMsgEnd is essentially an end position of the row in the buffer. And the only time when we can (and should) update it, is when the buffer is reset (that is, some data was requested from the buffer when it's full).

Yeah, I figured as much. FWIW the usage of _dataMsgEnd is slightly weird to me; previously it really was just an end position of the row in the buffer (because the entire row was always contained in the buffer). Now it can also be a point way beyond the end of the buffer, which actually has no real meaning in that case.

I think it should be possible to update it as we read more data into the buffer (or some similar mechanism), but I'm also fine with merging as-is if you don't feel like the extra effort is worth it (this is already a really nice improvement).