@vonzshik 2021-01-11
Backported to 5.0.2 via 945adc73ff6ee87ca579b7cc56935df741f5db4c