@vonzshik 2021-01-11
@roji I did some more benchmarking - in some cases sequential read might be slower by 2-3%. I'll create another issue and post my findings there.