@roji 2021-01-11
Note for the future: I think we shouldn't be releasing perf optimizations such as this in patch versions... We should strive to only fix bugs, and where it really makes sense, merge very low-risk features which have a lot of value.