@vonzshik 2021-01-11
> Note for the future: I think we shouldn't be releasing perf optimizations such as this in patch versions... We should strive to only fix bugs, and where it really makes sense, merge very low-risk features which have a lot of value.

Fully agree with this. Although, I do think this pr is relatively low-risk and does bring a lot of value. But if you're still against it, I'm willing to revert it.