@roji 2021-01-11
I do think we're better off reverting this - after all it really is a perf optimization, rather than a feature that could unblocked someone out there. The code around sequential/non-sequential reading has also proven to be trickier than it looks, so from a risk perspective this isn't great.

But I'll leave it up to your judgement... I just prefer to not have to worry about new bugs in patch releases...