@YohDeadfall 2021-01-11
Personally, I would like to have it in the upcoming release since it tunes perf a bit, but without drastic changes, and it doesn't look dangerous. It's clean and simple.