@YohDeadfall 2021-01-11
Not sure about it since there's not so much changes for 5.1, so I bet the next release is 6.0.