@roji 2021-01-11
We can indeed decide to release 5.1 at some point - this may be a good reason to hold off on breaking changes (since we don't want those in 6.0). The question is if we'll have enough interesting stuff to warrant a 5.1.